pyproject.toml
[project]
name = "campused"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "google-auth>=2.41.1",
    "psycopg2-binary>=2.9.10",
    "razorpay>=2.0.0",
    "requests>=2.32.5",
]
